requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.2  # Explicitly add this version
pydantic==2.6.4
web3==7.0.0b1
ipfshttpclient==0.8.0a2
opencv-python==4.9.0.80
tensorflow-cpu==2.16.1
pytest==8.1.1
httpx==0.27.0
pydantic-settings==2.2.1